MDEV-20844 RBR from binary(16) to inet6 fails with error 171: The event was corrupt, leading to illegal data being read
This patch changes the way how INET6 is packed to the RBR binary log,
from fixed length 16 bytes to BINARY(16) compatible variable length style
with trailing 0x00 byte compression.

This is to make INET6 fully compatible with BINARY(16),
so RBR replication works in this scenarios:

- Old master BINARY(16) -> New slave INET6
- New master INET6      -> Old slave BINARY(16)

A new class StringPack was added to share the code between
Field_string and Field_inet6.